deleted:    week4/include/dllist.h 	deleted:    week4/include/fields.h 	deleted:    week4/include/jrb.h 	deleted:    week4/include/jval.h 	deleted:    week4/lib/libfdr.a 	modified:   week4/phonebook/Makefile 	modified:   week4/phonebook/phonebook 	modified:   week4/phonebook/phonebook.c 	modified:   week4/phonebook/phonebook.h 	renamed:    week4/strsort/rockytop3.txt -> week4/phonebook/phonebook_mdf.txt 	modified:   week4/strsort/Makefile 	deleted:    week4/strsort/nsort.c 	deleted:    week4/strsort/nsort2.c 	deleted:    week4/strsort/nsort3.c 	deleted:    week4/strsort/strisort.c 	deleted:    week4/strsort/strrsort1.c 	deleted:    week4/strsort/strrsort2.c 	modified:   week4/strsort/strsort.c 	deleted:    week4/strsort/strusort.c 	deleted:    week4/strsort/testing.c/nsort.c 	deleted:    week4/strsort/testing.c/nsort2.c 	deleted:    week4/strsort/testing.c/nsort3.c 	deleted:    week5/main.c
	lib/
	week4/strsort/testingc/
	week5/
