--- week4/include/fields.h
@@ -1,63 +1,0 @@
-/*
-Libraries for fields, doubly-linked lists and red-black trees.
-Copyright (C) 2001 James S. Plank
-
-This library is free software; you can redistribute it and/or
-modify it under the terms of the GNU Lesser General Public
-License as published by the Free Software Foundation; either
-version 2.1 of the License, or (at your option) any later version.
-
-This library is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-Lesser General Public License for more details.
-
-You should have received a copy of the GNU Lesser General Public
-License along with this library; if not, write to the Free Software
-Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
-
----------------------------------------------------------------------------
-Please see http://www.cs.utk.edu/~plank/plank/classes/cs360/360/notes/Libfdr/
-for instruction on how to use this library.
-
-Jim Plank
-plank@cs.utk.edu
-http://www.cs.utk.edu/~plank
-
-Associate Professor
-Department of Computer Science
-University of Tennessee
-203 Claxton Complex
-1122 Volunteer Blvd.
-Knoxville, TN 37996-3450
-
-     865-974-4397
-Fax: 865-974-4404
- */
-#ifndef _FIELDS_ 
-#define _FIELDS_ 
-
-/* The fields library -- making input processing easier */
-
-#include <stdio.h>
-#define MAXLEN 1001
-#define MAXFIELDS 1000
-
-typedef struct inputstruct {
-  char *name;               /* File name */
-  FILE *f;                  /* File descriptor */
-  int line;                 /* Line number */
-  char text1[MAXLEN];       /* The line */
-  char text2[MAXLEN];       /* Working -- contains fields */
-  int NF;                   /* Number of fields */
-  char *fields[MAXFIELDS];  /* Pointers to fields */
-  int file;                 /* 1 for file, 0 for popen */
-} *IS;
-
-extern IS new_inputstruct(/* FILENAME -- NULL for stdin */);
-extern IS pipe_inputstruct(/* COMMAND -- NULL for stdin */);
-extern int get_line(/* IS */); /* returns NF, or -1 on EOF.  Does not
-                                  close the file */
-extern void jettison_inputstruct(/* IS */);  /* frees the IS and fcloses 
-                                                the file */
-#endif
