+++ week4/phonebook/Makefile
@@ -1,28 +1,58 @@
-EXECUTABLES = phonebook
-SRCS= phonebook.c
+# project name (generate executable with this name)
+TARGET = phonebook
 
-IDIR =../include
-CFLAGS= -g -Wall -I$(IDIR)
 CC=gcc
+# compiling flags here
+CFLAGS= -g -Wall
 
-ODIR=obj
-LDIR =../lib
-LIBS = libfdr.a
+IDIR = ../../lib/include
+# define any directories containing header file other then /usr/include
+INCLUDES = -I$(IDIR)
 
-DEPS = fields.h jrb.h jval.h dllist.h
-DEPS = $(patsubst %,$(IDIR)/%,$(_DEPS))
+# define any libraries to link into executable:
+# if I want to link in libraries
+# (libx.so or libx.a) I use the -llibname (-lx -lx) = (-lx)
+# libfdr => -lfdr
+LIBS = -lfdr 
+# LIBS = -lm
 
-_OBJ= $(SRCS:.c=.o)
-OBJ = $(patsubst %,$(ODIR)/%,$(_OBJ))
+# define any directories containing implementation of the file in INCLUDES
+# ../..lib means up 2 level then search for /lib
+LDIR = ../../lib/libs
+LFLAGS = -L$(LDIR)
+
+# This uses Suffix Replacement within a macro:
+#   $(name:string1=string2)
+# For each word in 'name' replace 'string1' with 'string2'
+# Below we are replacing the suffix .c of all words in the macro SRCS
+# with the .o suffix
+
+SRCS = $(wildcard *.c)
+ODIR = obj
+_OBJS = $(SRCS:.c=.o)
+OBJS = $(patsubst %,$(ODIR)/%,$(_OBJS))
+
+#_DEPS = fields.h jrb.h jval.h dllist.h
+#DEPS = $(patsubst %,$(IDIR)/%,$(_DEPS))
+
+
+.PHONY: depend clean
 #action
-all: $(EXECUTABLES)
+all: 	$(TARGET) 
+
+# This uses Suffix Replacement within a macro:
+#   $@ = left-part of ":" ( == $(TARGET))
+# 
+$(TARGET): $(OBJS)
+	$(CC) $(CFLAGS) $(INCLUDES) -o $@ $^ $(LFLAGS) $(LIBS)
+	@echo "Linking complete!"
 
 $(ODIR)/%.o: %.c $(DEPS)
-	$(CC) -c -o $@ $< $(CFLAGS)
-#change here
-phonebook: $(ODIR)/phonebook.o
-	$(CC) -o $@ $^ $(CFLAGS) $(LDIR)/$(LIBS)
-.PHONY: clean
+	$(CC) $(CFLAGS) $(INCLUDES) -c $< -o $@
+	@echo "Compiled "$<" successfully!"
 
 clean:
-	rm -f $(ODIR)/*.o core $(EXECUTABLES)
+	rm -f $(ODIR)/*.o *~ $(TARGET)
+	@echo "Cleanup complete!"
+#depend: $(SRCS)
+#        makedepend $(INCLUDES) $^
