--- week4/phonebook/phonebook_mdf.txt
@@ -1,23 +1,0 @@
-Bùi Thị Kim Quyên	(607)538-4992
-Cao Nguyễn Huỳnh Minh	(360)631-7949
-Cao Thị Minh Hiền	(334)448-8593
-Châu Thị Kim Anh	(570)588-2431
-Doãn Phan Trung Hải	(469)764-5493
-Dương Hoài Phương	(563)487-1451
-Lê Minh Minh Vương	(254)326-8583
-Mai Nguyễn Khánh Vân	(210)467-1665
-Nguyen Thi Thanh Bích	(419)769-7921
-Nguyễn Hoàng Thế Vinh	(475)999-9996
-Nguyễn Minh Châu	(469)256-2969
-Nguyễn Quỳnh Trân	(601)872-4415
-Nguyễn Thiện Văn Toàn	(213)375-9716
-Nguyễn Thị Vân Anh	(401)305-2672
-Nguyễn Võ An Thiện	(253)891-9848
-Phan Bùi Vinh Bính	(682)730-8447
-Phan Huỳnh Ngọc Dung	(574)248-7337
-Ta thị Thanh Tuyền	(662)307-4502
-Trương Nguyễn Gia Mẫn	(262)278-3444
-Võ Hoàng Minh Thư	(551)800-6879
-Võ Thị Tuyết Vân	(617)793-1182
-Vương Thị Thu Hiền	(616)988-4178
-Đoàn Hoàng Mỹ Xuân	(574)658-2494
