--- week5/main.c
@@ -1,8 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-
-int main()
-{
-    
-    return 0;
-}
